Rename CategoryItem Object
