+++ Growing-Traveler/Growing-Traveler/Model/StudyGoal/StudyGoal.swift
@@ -1,0 +1,41 @@
+//
+//  StudyItem.swift
+//  Growing-Traveler
+//
+//  Created by Jenny Hung on 2022/4/10.
+//
+
+import Foundation
+
+struct StudyGoal: Codable {
+    
+    var title: String
+    
+    var category: CategoryItem
+    
+    var studyPeriod: [StudyPeriod]
+    
+    var studyItems: [StudyItem]
+    
+    var createTime: TimeInterval
+}
+
+struct StudyPeriod: Codable {
+    
+    var startTime: Date
+    
+    var endTime: Date
+    
+}
+
+struct StudyItem: Codable {
+    
+    var itemTitle: String
+    
+    var studyTime: Int
+    
+    var content: String
+    
+    var isCompleted: Bool
+    
+}
