--- Growing-Traveler/Growing-Traveler/Model/StudyGoal/StudyItem.swift
@@ -1,20 +1,0 @@
-//
-//  StudyItem.swift
-//  Growing-Traveler
-//
-//  Created by Jenny Hung on 2022/4/10.
-//
-
-import Foundation
-
-struct StudyItem: Codable {
-    
-    var itemTitle: String
-    
-    var studyTime: Int
-    
-    var content: String
-    
-    var isCompleted: Bool
-    
-}
